add tree node split function

+++ include/data.h
@@ -1,7 +1,6 @@
 #ifndef _TREE_H_
 #define _TREE_H_
 
-#include "vector.h"
 
 namespace gbrt_cy
 {
@@ -15,7 +14,7 @@ struct Config {
 class Data {
 public:
     uint64_t sample_size;
-    uint64_t feature_size;
+    uint32_t feature_size;
     double** samples;
     double* residual;
     double* predict;
